fix: remove header passing for fetching subgraphs
