+++ lib/composer.js
@@ -108,8 +108,8 @@ class Composer {
     return printSchema(buildClientSchema(this.toSchema()))
   }
 
-  async compose (options) {
-    this.schemas = await this.fetchSubgraphSchemas(options)
+  async compose () {
+    this.schemas = await this.fetchSubgraphSchemas()
 
     for (let i = 0; i < this.schemas.length; ++i) {
       this.mergeSchema(this.schemas[i])
@@ -241,11 +241,11 @@ class Composer {
     }
   }
 
-  async fetchSubgraphSchemas (options) {
+  async fetchSubgraphSchemas () {
     const subgraphs = Array.from(this.subgraphs.values())
 
-    const requests = subgraphs.map(({ server }) => {
-      return fetchSubgraphSchema({ server, ...options })
+    const requests = subgraphs.map((subgraph) => {
+      return fetchSubgraphSchema(subgraph.server)
     })
 
     const responses = await Promise.allSettled(requests)
