+++ lib/index.js
@@ -1,10 +1,10 @@
 'use strict'
 const { Composer } = require('./composer')
 
-async function compose (config, options) {
+async function compose (config) {
   const composer = new Composer(config)
 
-  await composer.compose(options)
+  await composer.compose()
   return composer
 }
