+++ lib/network.js
@@ -14,16 +14,16 @@ function graphqlRequest ({ url, headers, query, variables }) {
   })
 }
 
-async function fetchSubgraphSchema ({ server, headers }) {
+async function fetchSubgraphSchema (server) {
   const composeEndpoint = `${server.host}${server.composeEndpoint}`
   const endpoints = [composeEndpoint]
-  let response = await request(composeEndpoint, { headers })
+  let response = await request(composeEndpoint)
 
   if (response.statusCode !== 200) {
     // fallback to graphqlEndpoint sending introspection query
     const graphqlEndpoint = `${server.host}${server.graphqlEndpoint}`
     endpoints.push(graphqlEndpoint)
-    const gqlResponse = await graphqlRequest({ url: graphqlEndpoint, query: introspectionQuery, headers })
+    const gqlResponse = await graphqlRequest({ url: graphqlEndpoint, query: introspectionQuery })
 
     if (gqlResponse.statusCode !== 200) {
       const msg = `Unable to get schema from ${composeEndpoint} (response ${gqlResponse.statusCode}) nor ${graphqlEndpoint} (response ${gqlResponse.statusCode})`
